Expand on the testing around watching a show.
Change the working to startWatching instead of watch. Also expanded
on what the Series Source will be required to do. Added the new
requirement to the Test Series DB.

Included the start of stopWatching, but it's not tested, and therefore
not working.

Tried to convert the tests slightly to a more BDD design.