Add a very basic python UI.
This UI connects and will allow you run commands on the client like
client.getWatching, or client.search("show").

The results are not cleaned at all. Also the commands couldn't easily
be added to the buildins because the __dict__ on the objects didn't
contain the commands. We can look in to this later.

It might be that we'll need to create a python UI class that will make
the calls and clean the output. This could be added to the interact
to make a nice list of commands.